Refactor config.js; Move random-data import (#163)
1. config.js exports a function that creates an object literal. Some of these values could easily just be exportable string constants.
2. I've moved dataSource into its own file, and imported random data within it. This is because it's only called in `src/index.js`, so it would not be imported inside `src/components/App.js`, which means that the random data generator code would not be included in the exported package, which should reduce file-sizes slightly for projects that import Kedro-Viz. This is great because the random-data generator is only really useful in development/testing, and isn't that useful for package consumers
