+++ src/config.test.js
@@ -1,43 +1,19 @@
-import config from './config';
+import { dataPath, localStorageName } from './config';
 
 describe('config', () => {
-  const OLD_ENV = process.env;
+  describe('dataPath', () => {
+    it('should return the name of a json file', () => {
+      expect(dataPath).toEqual(expect.stringContaining('.json'));
+    });
 
-  beforeEach(() => {
-    jest.resetModules();
-    process.env = { ...OLD_ENV };
-    delete process.env.REACT_APP_DATA_SOURCE;
-  });
-
-  afterEach(() => {
-    process.env = OLD_ENV;
-  });
-
-  it('should return a properly-formatted object', () => {
-    expect(config()).toEqual({
-      dataPath: expect.stringContaining('.json'),
-      dataSource: expect.stringMatching(/random|json/),
-      localStorageName: expect.stringContaining('KedroViz')
+    it('should return a relative path', () => {
+      expect(dataPath.substr(0, 2)).toEqual('./');
     });
   });
 
-  it('should return a relative dataPath', () => {
-    expect(config().dataPath.substr(0, 2)).toEqual('./');
-  });
-
-  it('should return "json" as the datasource if undefined', () => {
-    expect(config().dataSource).toBe('json');
-  });
-
-  it('should throw an error if the given source is unknown', () => {
-    process.env.REACT_APP_DATA_SOURCE = 'qwertyuiop';
-    expect(() => {
-      config().dataSource();
-    }).toThrow();
-  });
-
-  it('should return the given datasource if set', () => {
-    process.env.REACT_APP_DATA_SOURCE = 'random';
-    expect(config().dataSource).toBe('random');
+  describe('localStorageName', () => {
+    it('should contain KedroViz', () => {
+      expect(localStorageName).toEqual(expect.stringContaining('KedroViz'));
+    });
   });
 });
