+++ src/store/initial-state.test.js
@@ -1,22 +1,35 @@
 import getInitialState, { getPipelineData } from './initial-state';
 import { saveState } from './helpers';
+import getRandomPipeline from '../utils/random-data';
 import loremIpsum from '../utils/data/lorem-ipsum.mock';
 import animals from '../utils/data/animals.mock';
 import demo from '../utils/data/demo.mock';
+import layers from '../utils/data/layers.mock';
 
 describe('getPipelineData', () => {
   it('returns the correct dataset when passed a dataset string', () => {
-    expect(getPipelineData('random')).toMatchObject({
-      nodes: expect.any(Array),
-      edges: expect.any(Array),
-      tags: expect.any(Array)
-    });
-    expect(getPipelineData('random')).not.toMatchObject({
-      nodeName: expect.any(Object)
-    });
     expect(getPipelineData('lorem')).toEqual(loremIpsum);
     expect(getPipelineData('animals')).toEqual(animals);
     expect(getPipelineData('demo')).toEqual(demo);
+    expect(getPipelineData('layers')).toEqual(layers);
+  });
+
+  it("returns null when passed 'json'", () => {
+    expect(getPipelineData('json')).toEqual(null);
+  });
+
+  it('returns original data', () => {
+    const randomPipeline = getRandomPipeline();
+    expect(getPipelineData(randomPipeline)).toEqual(randomPipeline);
+  });
+
+  it('throws an error when passed random', () => {
+    expect(() => getPipelineData('random')).toThrow();
+  });
+
+  it('throws an error when passed null or undefined', () => {
+    expect(() => getPipelineData(null)).toThrow();
+    expect(() => getPipelineData(undefined)).toThrow();
   });
 });
