+++ src/store/load-data.js
@@ -1,11 +1,10 @@
 import { json } from 'd3-fetch';
-import config from '../config';
+import { dataPath } from '../config';
 
 /**
  * Asynchronously load, parse and format data from json file using D3
  */
 const loadJsonData = () => {
-  const { dataPath } = config();
   const fullPath = `/public${dataPath.substr(1)}`;
 
   return json(dataPath).catch(() => {
